+++ app/views/playground.vue
@@ -139,20 +139,20 @@
 					>
 						<template v-slot:system="slot">
 							<g v-if="enabledChromatic">
-								<g class="staff" v-for="(staff, iii) of slot.row.staves" :key="iii"
+								<g class="staff" v-for="(staff, iii) of slot.system.staves" :key="iii"
 									:transform="`translate(${staff.x}, ${staff.y})`"
 								>
 									<g v-if="chromaticMode === 'symbols'" class="locator">
 										<rect class="head" :x="0" :y="-2" :width="staff.headWidth" :height="4" />
 										<circle />
-										<line v-if="Number.isFinite(staff.top)" :x1="0" :y1="staff.top" :x2="slot.row.width" :y2="staff.top" />
+										<line v-if="Number.isFinite(staff.top)" :x1="0" :y1="staff.top" :x2="slot.system.width" :y2="staff.top" />
 										<g class="measure" v-for="(measure, i4) of staff.measures" :key="i4" :class="measure.class">
 											<rect :x="measure.lineX" :y="-2" :width="measure.noteRange.end - measure.lineX" :height="4"/>
 											<text :x="measure.headX">'{{measure.index}}</text>
 										</g>
 									</g>
-									<g v-if="chromaticMode === 'pitch' && pitchContextMarks[slot.row.index] && pitchContextMarks[slot.row.index][iii]" class="pitch-context">
-										<g v-for="(item, i4) of pitchContextMarks[slot.row.index][iii]" :key="i4" :transform="`translate(${item.x}, 0)`">
+									<g v-if="chromaticMode === 'pitch' && pitchContextMarks[slot.system.index] && pitchContextMarks[slot.system.index][iii]" class="pitch-context">
+										<g v-for="(item, i4) of pitchContextMarks[slot.system.index][iii]" :key="i4" :transform="`translate(${item.x}, 0)`">
 											<line class="base-line" x1="0" x2="0" y1="-3" y2="3" />
 											<g transform="translate(0.2, 0.5)">
 												<g v-for="(pitch, i5) of item.names[0]" :key="i5" class="pitch"
@@ -544,17 +544,17 @@
 					const context = item.context;
 					const yToName = y => ({y, alter: context.alterOnNote(context.yToNote(y)), name: context.yToPitchName(y)});
 
-					const {row, x} = this.scheduler.lookupPosition(item.tick);
+					const {system, x} = this.scheduler.lookupPosition(item.tick);
 					const names = [
 						[-2, -1, 0, 1, 2].map(yToName),
 						[-1.5, -0.5, 0.5, 1.5].map(yToName),
 					];
 
-					return {row, x, names};
-				})).reduce((result, table, staff) => (table.forEach(({row, ...item}) => {
-					result[row] = result[row] || [];
-					result[row][staff] = result[row][staff] || [];
-					result[row][staff].push(item);
+					return {system, x, names};
+				})).reduce((result, table, staff) => (table.forEach(({system, ...item}) => {
+					result[system] = result[system] || [];
+					result[system][staff] = result[system][staff] || [];
+					result[system][staff].push(item);
 				}), result), []);
 			},
 		},